# openapi/device-api.yaml
openapi: 3.1.0
info:
  title: Smart Access Control — Device API
  version: 1.0.0
  summary: The contract between an ESP32-S3 access panel and the backend.
  description: |
    Every path here is appended directly to the device's configured
    `api_base_url`, which already carries the version prefix:

        https://example.com/api/v1  +  /health

    ### Constraints this contract is shaped by

    These come from the shipped firmware and are not negotiable:

    * **The keypad is a 4x3 matrix — digits, `*` and `#`.** There is no way to
      type a letter, so every code the backend issues is six *digits*. An
      alphanumeric code is physically unenterable at the door.
    * **The device has no real-time clock.** Its clock reads 1970 until NTP
      lands. The server assigns every authoritative timestamp; a device-supplied
      wall clock is never trusted. `uptime_ms` orders events within one batch
      and nothing more.
    * **Fingerprint templates live in the sensor's own flash.** Slot ids are
      device-local integers `0 … capacity-1` and survive reboots and reflashes.
      The backend records who owns a slot; it can never move a template.
    * **The device must keep working while offline.** Events are queued in
      flash and flushed on reconnect, so this API accepts batched, replayed
      submissions with an idempotency key.
    * **Phone numbers are capped at 14 characters** in device storage. Numbers
...
      refused.

...

...
       console.

  contact:
...
  license:
...
servers:
...
security:
  - DeviceId: []
    ApiKey: []
    ApiSecret: []
...
paths:
...
  /otp-requests:
    post:
      tags: [OTP]
      operationId: requestOtp
      summary: Issue a one-time password for a phone number
      description: |
        **The response never contains the code.** The backend delivers it over
        SMS and e-mail and the device submits what the user typed to
        `POST /otp-verifications`.

        `valid: false` means the number is not known to this device's tenant;
        the panel shows "Phone number not found" and lets the user retry. A
        suspended member gives exactly the same answer as an unknown number —
        a door is not a place to enumerate people.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                device_id: { $ref: '#/components/schemas/DeviceIdClaim' }
                phone: { type: string, examples: ['+94771234567'] }
      responses:
        '200':
          description: The outcome of the lookup.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    title: Issued
                    required: [valid, otp_request_id, expires_in]
                    properties:
                      valid: { type: boolean, const: true }
                      otp_request_id: { type: string, format: uuid }
                      expires_in:
                        type: integer
                        description: Seconds until the code expires.
                        examples: [90]
                      otp:
                        type: string
                        pattern: '^\d{6}$'
                        description: |
                          **Legacy compatibility only**, present when
                          `access.otp.return_code_to_device` is enabled. This
                          hands the plaintext code to anyone holding the device
                          credential and is off by default.
                  - type: object
                    title: Not found
                    required: [valid]
                    properties:
                      valid: { type: boolean, const: false }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /otp-verifications:
    post:
      tags: [OTP]
      operationId: verifyOtp
      summary: Verify the code the user typed
      description: |
        Always 200 — a wrong code is an answer, not a transport failure, and
        the panel branches on `verified`. Every outcome is written to the
        access log.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp_request_id, code]
              properties:
                device_id: { $ref: '#/components/schemas/DeviceIdClaim' }
                otp_request_id: { type: string, format: uuid }
                code:
                  type: string
                  pattern: '^\d{6}$'
                  examples: ['482915']
      responses:
        '200':
          description: The verification decision.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    title: Verified
                    properties:
                      verified: { type: boolean, const: true }
                      member_id: { type: string, format: uuid }
                      event_id: { type: string, format: uuid }
                  - type: object
                    title: Refused
                    properties:
                      verified: { type: boolean, const: false }
                      reason:
                        type: string
                        enum: [otp_mismatch, otp_expired, member_suspended, outside_schedule]
                      attempts_left: { type: integer, examples: [2] }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    DeviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: |
        The identifier the firmware generates on first boot (`SMA_4821`).
        **Device-asserted**, so the backend treats it as a claim: it is checked
        against the device the credential belongs to and never trusted alone.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    ApiSecret:
      type: apiKey
      in: header
      name: X-API-Secret
      description: |
        Compared in constant time against a stored digest. Two live credentials
        per device are permitted so a rotation overlaps instead of leaving a
        gap where the panel cannot reach the backend.
...
  schemas:
    DeviceIdClaim:
      type: string
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
      description: |
        Optional echo of the device's own id. If present it must match the
        authenticated device; a mismatch is a 403, because it means either a
        misconfigured panel or a credential being used to write another
        device's history.
      examples: ['SMA_4821']
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: The contract. Branch on this, not on the message.
              examples: ['otp_expired']
            message:
              type: string
              description: For whoever reads the logs afterwards.
            fields:
              type: object
              description: Field-level detail on a validation failure.
              additionalProperties:
                type: array
                items: { type: string }
            retry_after:
              type: integer
              description: Seconds, on a rate-limit response.
  responses:
    Unauthorized:
      description: |
        Bad credentials. Deliberately identical for every failure mode — which
        of the three header values was wrong is not something a caller gets to
        learn by probing.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: unauthorized
              message: Invalid device credentials.
...
    RateLimited:
      description: |
        Per-device limit exceeded. Keyed on the authenticated device rather
        than the IP: a whole site's panels commonly share one public address,
        and one compromised device must not throttle its neighbours.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }